Examples/FunctionApps/GlupFunctionApp/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
tensorflow==2.5.0
numpy==1.19.2
h5py==3.1.0
six==1.15.0
scipy==1.4.1
Pillow==8.2.0
opencv-python==4.5.2.54